Disable FORTIFY_SOURCE
glibc has added a feature which enabled compile-time and run-time checking for
buffer overflows. However this requires glibc 2.15, running on a previous
glibc on an embedded system when cross compiled may give the following error:

  version `GLIBC_2.15' not found

which is to do with the __fdelt_chk function not being available.

Disabling this feature is a workaround which we will use for now.