Improve handling of USB enumeration on reset
The old logic assumed that the device would re-enumerate either at the same
address or the maximum device address on the bus plus one. This logic didn't
handle the case where the device address reached 127 and wrapped.

Instead, build a list of which addresses on the bus are maxtouch devices,
which allows us to detect the new one that has arrived.

Signed-off-by: Nick Dyer <nick.dyer@itdev.co.uk>