Read info from info block of .xcfg files
This reads family ID, variant, version, build, and the crc for
the info block from .xcfg files.  Vendor ID and Product ID are
still currently ignored as there is no location to store this
in the mxt_id_info struct (which is based on OBP_RAW V1 format).
There is no good way to get the matrix_xsize, matrix_ysize, or
num_objects values of the mxt_id_info struct, so these are
left unpopulated.